# faster parse_snapshot