--- a/src/HeapSnapshotParser.jl
+++ b/src/HeapSnapshotParser.jl
@@ -4,26 +4,28 @@ using JSON
 using LightGraphs
 using StructEquality
 
-@struct_hash_equal Base.@kwdef struct Node
-    kind::Symbol
-    type::String
-    id::Int
-    num_edges::Int
-    self_size::Int
+@struct_hash_equal Base.@kwdef mutable struct Node
+    kind::Union{Symbol,Nothing} = nothing
+    type::Union{String,Nothing} = nothing
+    id::Union{Int,Nothing} = nothing
+    num_edges::Union{Int,Nothing} = nothing
+    self_size::Union{Int,Nothing} = nothing
 
     # index into snapshot.edges
     # would be an array of indexes, but Julia doesn't
     # support types which refer to each other :facepalm:
     # https://github.com/JuliaLang/julia/issues/269
-    out_edge_indexes::Array{Int}
+    out_edge_indexes::Array{Int} = Int[]
 end
 
-@struct_hash_equal Base.@kwdef struct Edge
-    kind::Symbol
-    name::String
+dummy_node = Node()
 
-    from::Node
-    to::Node
+@struct_hash_equal Base.@kwdef mutable struct Edge
+    kind::Symbol = :null
+    name::String = ""
+
+    from::Node = dummy_node
+    to::Node = dummy_node
 end
 
 Base.@kwdef struct HeapSnapshot
@@ -37,8 +39,11 @@ NUM_NODE_FIELDS = length(NODE_FIElDS)
 EDGE_FIELDS = ["type", "name_or_index", "to_node"]
 NUM_EDGE_FIELDS = length(EDGE_FIELDS)
 
-function parse_snapshot(input::IOStream)::HeapSnapshot
-    parsed = JSON.parse(input)
+function parse_snapshot(file::String)::HeapSnapshot
+
+    # JSON.parsefile is faster than JSON.parse because it uses mmap
+    parsed = JSON.parsefile(file)
+
     snapshot = HeapSnapshot(nodes=[], edges=[])
 
     node_kind_enum = parsed["snapshot"]["meta"]["node_types"][1]
@@ -47,29 +52,41 @@ function parse_snapshot(input::IOStream)::HeapSnapshot
     nodes = parsed["nodes"]
     strings = parsed["strings"]
     num_nodes = convert(Int, length(nodes)/NUM_NODE_FIELDS)
-    for node_idx = 0:(num_nodes-1)
+
+    all_num_edges = Vector{Int}(undef, num_nodes + 1)
+    all_num_edges[1] = 0
+    num_edges_accum = Vector{Int}(undef, num_nodes + 1)
+
+    append!(snapshot.nodes, map(_->Node(), 1:num_nodes))
+
+    Threads.@threads for node_idx = 0:(num_nodes-1)
         kind_key = nodes[node_idx*NUM_NODE_FIELDS + 1]
         name_key = nodes[node_idx*NUM_NODE_FIELDS + 2]
         id = nodes[node_idx*NUM_NODE_FIELDS + 3]
         self_size = nodes[node_idx*NUM_NODE_FIELDS + 4]
         num_edges = nodes[node_idx*NUM_NODE_FIELDS + 5]
 
-        node = Node(
-            kind=Symbol(node_kind_enum[kind_key + 1]),
-            type=strings[name_key + 1],
-            num_edges=num_edges,
-            self_size=self_size,
-            out_edge_indexes=[], # filled in below
-            id=id,
-        )
+        node = snapshot.nodes[node_idx + 1]
+        node.kind = Symbol(node_kind_enum[kind_key + 1])
+        node.type = strings[name_key + 1]
+        node.num_edges = num_edges
+        node.self_size = self_size
+        node.out_edge_indexes = [] # filled in below
+        node.id = id
 
-        push!(snapshot.nodes, node)
+        all_num_edges[node_idx + 2] = num_edges
+    end
+    cumsum!(num_edges_accum, all_num_edges)
+
+    for _ in 1:num_edges_accum[end]
+        push!(snapshot.edges, Edge())
     end
 
     edges = parsed["edges"]
-    edge_idx = 0
-    for from_node in snapshot.nodes
+    Threads.@threads for (node_i, from_node) in collect(enumerate(snapshot.nodes))
         for edge_num = 1:(from_node.num_edges)
+            edge_idx = num_edges_accum[node_i] + (edge_num - 1)
+
             kind_key = edges[edge_idx*NUM_EDGE_FIELDS + 1]
             name_key = edges[edge_idx*NUM_EDGE_FIELDS + 2]
             to_key = edges[edge_idx*NUM_EDGE_FIELDS + 3]
@@ -86,16 +103,13 @@ function parse_snapshot(input::IOStream)::HeapSnapshot
                 strings[name_key+1]
             end
 
-            edge = Edge(
-                kind=kind,
-                name=name,
-                from=from_node,
-                to=snapshot.nodes[to_node_idx],
-            )
+            edge = snapshot.edges[edge_idx + 1]
+            edge.kind = kind
+            edge.name = name
+            edge.from = from_node
+            edge.to = snapshot.nodes[to_node_idx]
 
-            push!(snapshot.edges, edge)
-            push!(from_node.out_edge_indexes, length(snapshot.edges))
-            edge_idx += 1
+            push!(from_node.out_edge_indexes, edge_idx)
         end
     end